requirements.txt
boto3
botocore
mysql-connector-python
pandas
openpyxl
python-dotenv